refactor(web): localize view models and push jobStatusSettled to Core
Keep Core DTOs in services; move TaskWithCoworker next to the tasks UI,
expose settlement on JobSummary/Job, and document the view-model rule.
