+++ apps/core/src/routes/v1/agents/[id]/jobs/post.test.ts
@@ -66,6 +66,7 @@ describe("POST /agents/{id}/jobs", () => {
       jobType: "FREE",
       status: "processing",
       credits: 0,
+      jobStatusSettled: false,
       onChainStatus: null,
       onChainTransactionHash: null,
       result: null,
