+++ apps/web/src/app/(app)/agents/[agentId]/jobs/components/__tests__/jobs-list-row.test.tsx
@@ -71,6 +71,7 @@ function createJob(
       image: userOverrides?.image ?? null,
     },
     ...rest,
+    jobStatusSettled: rest.jobStatusSettled ?? false,
   };
 }
