+++ apps/web/src/app/(app)/tasks/components/__tests__/task-activity.test.tsx
@@ -2,8 +2,8 @@ import { act, render, screen, waitFor } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import { beforeEach, describe, expect, it, vi } from "vitest";
 import { TaskActivitySection } from "@/app/tasks/components/task-activity";
+import type { TaskEvent } from "@/lib/clients/generated/core";
 import { Channel, TaskStatus } from "@/lib/clients/generated/core";
-import type { TaskEvent } from "@/lib/types/task";
 
 const {
   uploadTaskAttachmentMock,
