+++ apps/web/src/app/(app)/tasks/components/task-card.tsx
@@ -1,9 +1,9 @@
 "use client";
 
+import type { TaskWithCoworker } from "@/app/tasks/types/task-board";
 import { TaskScheduleDisplay } from "@/components/task-schedule-display";
 import { TaskStatus } from "@/lib/clients/generated/core";
 import type { TaskStatus as TaskStatusType } from "@/lib/types/core-dto";
-import type { TaskWithCoworker } from "@/lib/types/task";
 import { cn } from "@/lib/utils";
 import { hasActiveSchedule } from "@/lib/utils/task-schedule";
 import { TaskDetailLink } from "./task-detail-link";
