+++ apps/web/src/app/(app)/tasks/components/task-list-item.tsx
@@ -1,7 +1,7 @@
 "use client";
 
+import type { TaskWithCoworker } from "@/app/tasks/types/task-board";
 import { TaskStatus } from "@/lib/clients/generated/core";
-import type { TaskWithCoworker } from "@/lib/types/task";
 import { cn } from "@/lib/utils";
 import { TaskDetailLink } from "./task-detail-link";
 import type { DragHandleProps } from "./task-dnd";
