+++ apps/web/src/app/(app)/tasks/utils/tasks-column-page.ts
@@ -1,13 +1,16 @@
 import "server-only";
 
+import type {
+  KanbanColumnId,
+  TaskWithCoworker,
+} from "@/app/tasks/types/task-board";
+import { getColumnListQueryOptions } from "@/app/tasks/utils/task-column";
+import { mapTaskToTaskWithCoworker } from "@/app/tasks/utils/task-view-model";
 import type { TasksScope } from "@/app/tasks/utils/tasks-filters";
 import type { Coworker } from "@/lib/clients/generated/core";
 import { TaskStatus } from "@/lib/clients/generated/core";
 import { taskService } from "@/lib/services/task.service";
 import type { CoreAgentDto } from "@/lib/types/core-dto";
-import type { KanbanColumnId, TaskWithCoworker } from "@/lib/types/task";
-import { getColumnListQueryOptions } from "@/lib/utils/task-column";
-import { mapTaskToTaskWithCoworker } from "@/lib/utils/task-transformer";
 
 type ColumnCursor = string | null;
