+++ apps/web/src/components/jobs/job-details/__tests__/job-details-view.test.tsx
@@ -160,6 +160,7 @@ function createJob(overrides?: Partial<Job>): Job {
       overrideLegalOther: null,
     },
     ...overrides,
+    jobStatusSettled: overrides?.jobStatusSettled ?? false,
   };
 }
