+++ apps/web/src/components/jobs/job-details/__tests__/job-meta-details.test.tsx
@@ -71,6 +71,7 @@ function createJob(overrides: Partial<Job> = {}): Job {
       name: "Agent",
     },
     ...overrides,
+    jobStatusSettled: overrides?.jobStatusSettled ?? false,
   };
 }
