+++ apps/web/src/components/jobs/job-details/__tests__/job-share-modal.test.tsx
@@ -102,6 +102,7 @@ function createJob(overrides: Partial<Job>): Job {
       name: "Agent",
     },
     ...overrides,
+    jobStatusSettled: overrides?.jobStatusSettled ?? false,
   };
 }
