+++ apps/web/src/lib/helpers/job.ts
@@ -1,40 +1,15 @@
 import type { JobStatusData } from "@/lib/ably";
 import type { Job, JobSummary } from "@/lib/clients/generated/core";
-import { JobType } from "@/lib/clients/generated/core";
-
-export type CoreJobListItem = JobSummary & {
-  jobStatusSettled?: boolean;
-};
-
-type JobWithSettlementFields = Pick<
-  JobSummary,
-  "jobType" | "completedAt" | "externalDisputeUnlockTime"
->;
-
-export function isJobSettled(job: JobWithSettlementFields): boolean {
-  switch (job.jobType) {
-    case JobType.FREE:
-      return job.completedAt != null;
-    case JobType.PAID:
-      return job.externalDisputeUnlockTime
-        ? new Date() > job.externalDisputeUnlockTime
-        : false;
-    default:
-      return false;
-  }
-}
 
 /**
- * Get the job status data for the job which is used on sidebar job status indicator
- * and used by ably to update the job status in real time.
+ * Job status payload for the sidebar indicator and Ably realtime updates.
+ * Prefer `jobStatusSettled` from Core `JobSummary` / `Job` DTOs.
  */
-export function getJobStatusData(
-  job: JobSummary | Job | CoreJobListItem,
-): JobStatusData {
+export function getJobStatusData(job: JobSummary | Job): JobStatusData {
   return {
     jobId: job.id,
     jobStatus: job.status,
-    jobStatusSettled: isJobSettled(job),
+    jobStatusSettled: job.jobStatusSettled,
   };
 }
